coding/requirements.txt
Flask
flask-socketio
paho-mqtt
eventlet